Filter archive_sites based on prefix_frozen
This contains the prefix value that was set in macports.conf. If a port
changes prefix to something different, that doesn't disqualify it from
being installed from an archive, since the change is made on the build
server as well.

(cherry picked from commit d840f251bed8f6779048efdac97c9d9824fa3320)