mportinit: Disable running unknown cmds in $SHELL
Tcl defaults to running unknown commands in a shell automatically. This
can inadvertently lead to running things in the shell instead of using
a Tcl builtin, e.g. due to a typo.

It does not seem like we are currently using this functionality, and we
really shouldn't, so disable it on startup to ensure it is not used.

(cherry picked from commit fbc1d79b3728164962cb02e7f7d2d723d94421b0)