Adjust lint --niptick on patchfiles to allow *.diff or *.patch
The current behavior is to only allow patch-*.diff patchfiles.

Ryan Schmidt's comment on ticket 59695:
The lint warning was originally added to discourage users from adding
patchfiles that had neither the .diff nor the .patch extension, since
that resulted in incorrect syntax coloring in text editors that
determine the filetype from the extension.

closes https://trac.macports.org/ticket/59695