macports1.0: Selfupdate using HTTP
Refactor the selfupdate package into smaller functions and switch to
a control flow of:

 1. checking for newer versions by downloading a single URL
 2. downloading the new version using HTTP when available
 3. verifying the signature using EdDSA with ed25519 elliptic curve
    crypto with the help of OpenBSD's signify(1)
 4. automatically re-executing selfupdate for the sync step with the new
    base version when base was updated

This allows us to automatically fall back to other mirrors should our
main mirror not be available. We did not have this functionality
available for our previous rsync-based mechanism.

Additionally, the use of our standard curl wrapper enables displaying
a progress bar for the download and automatically uses any proxies that
might already be configured for HTTP.

Since we currently do not have a standard mechanism to sign the source
code tarballs uploaded to our distfiles server and github releases page
(we only sign the tarball pushed to rsync), I've taken the liberty to
introduce a new modern signature scheme for this channel – if we have to
touch this anyway, we might as well use modern tools and algorithms.

Previously, we would instruct users to re-run selfupdate manually when
MacPorts base was updated, because Portfiles might already require the
new base version to index correctly. This step has now been automated
using an idea that lives on the migration branch, where the port.tcl
client automatically re-executes itself in this situation.

This patch also adds support for a list of mirrors specified in the
ports tree at _resources/port1.0/fetch/base_mirror_sites.list, so that
we can add and remove mirrors without releasing a new version of
MacPorts base. Base itself comes preconfigured with the GitHub releases
URL if this list is not available.

See: https://trac.macports.org/ticket/60608