tests: Fix curl isnewer test broken by braeburn update
Updating braeburn to a subversion version >= 1.10.0 broke this test,
because the subversion backend no longer delivers the Last-Modified
header.

Switch to a URL under our control that does and, adjust the other tests
as required, update the timestamps to check in the 'curl isnewer'
testcases, and drop unused variables.

Ideally, we wouldn't be relying on a remote server for this test and
start a local one, but Apple doesn't ship /usr/bin/python3 on modern
systems, we should probably not rely on /usr/bin/python being around for
very much longer, and I'm not aware of any other quick-and-dirty
webservers we could start for the test short of writing our own.

Closes: https://trac.macports.org/ticket/60630