aclocal.m4: Run distclean before config in subdirs
Some of our vendored dependencies are not reconfigure-safe without
a distclean when changing, for example, --prefix. Specifically, this
applies to Tcl, which does not recompile some of the files that contain
the hardcoded prefix.

This leads to incorrect behavior when such a reconfigured MacPorts build
directory is installed – the bundled Tcl interpreter may end up loading
its packages from the previously configured --prefix.

Avoid this by always running 'make distclean' in the build directory
before running configure, but ignoring any errors, e.g. if the tree was
not yet configured, or no distclean target exists in the Makefile.

See https://github.com/macports/macports-base/pull/79 and
https://github.com/macports/macports-base/pull/80 for previous
discussion of the issue.