Don't add extra quoting to startupitems key values
This allows multi-command init, start, stop, and restart values to work
and allows executable values with arguments to work. This gives the
startupitems option feature parity with the older startupitem options.

Unfortunately it is also a change in syntax for existing users of the
init, start, stop, restart, and executable keys of the startupitems
option if the values of those keys contain whitespace. An additional
level of quoting is now required.

(cherry picked from commit 77161291d15398ce5894ef0db771045fbe0a62b9)