port1.0: Ensure adduser drops privs in all cases
adduser and addgroup (invoked by the add_users statement in
handle_add_users) elevates its privileges, but does not always drop them
again. Specifically, this could happen if a user or group already
existed, in which case adduser did elevate to root, but not drop back to
the macportsuser again.

This can cause files to be created with incorrect permissions, which can
then cause permissions problems in subsequent phases.

Closes: https://trac.macports.org/ticket/50918