Changelog: Document 2.6 changes on master
We shouldn't have diverging changelogs for master and the release
branches, so the Changelog entry I recently pushed to release-2.6 was
wrong and should have been made here on master in the first place.

Correct this so it can be cherry-picked to release-2.6.

(cherry picked from commit 059cf65170513b30e08b985d89bad3911666415b)