Always pass fs-traverse a list of paths
fs-traverse takes a list of paths to traverse. Passing just a path works
if the path doesn't contain any whitespace but fails if it does, so pass
the path in a list which always works.

This fixes, for example, chown failing when passed a path that contains
a space.

(cherry picked from commit 52609c9cce9da9d450aa3ccccc08b48b37c49788)