port1.0: configure: Add check for implicit decls
On macOS 11, implicit function declarations are now errors by default.
This may affect configure checks and actually cause software to compile
differently without us knowing.

Add a post-configure check that scans any config.log files in
${configure.dir} for indicators of such problems and warn about them, so
they will not go unnoticed. Extract the list of functions that were
found to be undeclared and check them against a whitelist of glob
expressions initialized from a file in

  _resources/port1.0/checks/implicit_function_declaration/macosx${configure.sdk_version}.sdk.list

so that functions that are legitimately undeclared on macOS can be
whitelisted globally. Note that the whitelist in

  configure.checks.implicit_function_declaration.whitelist

can also be edited per-Portfile, should that become necessary.