ci: azure: Fix deployment condition
The source branch is in the Build.SourceBranch variable, which is not
the same as variables.Build.SourceBranch - the former expects a variable
named 'Build.SourceBranch', the latter expects 'Build' to be a dict that
contains the key 'SourceBranch'.

Apparently the former is the correct approach, and the latter just
returns null.