ci: Add GitHub Actions, Azure Pipelines CI
Drop the Travis builds and move to GitHub Actions for 10.15 and 11.
GitHub Actions does not provide 10.14 images, so add Azure Pipelines for
10.14 as long as they still provide it.

These new CI steps perform configuration, build, installation and test
in separate steps, which gives us (a) better reporting, and (b) a better
user experience.

Additionally, this obsoletes the current re-deployment procedure of the
macports-ports CI environment that involved merging newly released tags
into a special branch; this step is now replaced by two conditional
steps in the CI that will run for pushed tags and perform the necessary
steps automatically.