darwintrace: Support arbitrary path lengths
darwintrace.dylib did occasionally crash in __darwintrace_is_in_sandbox
due to stack canary violations, because the reimplementation of realpath
and splitting into path components assumed that paths will not be longer
than MAXPATHLEN. In practice, this assumption was wrong, and many build
systems will invoke file system operations with paths longer than that.

Switch to dynamically allocated strings and arrays and resize the
buffers as necessary. This also simplifies the code significantly and
fixes a number of corner cases that were previously not handled
correctly, such as walking up past the root directory or multiple
slashes in a row.

With these changes, quassel builds successfully in trace mode – without
them, the build abort(3)s.