support compiler.c_standard 2017
This doesn't do a whole lot since this revision of the standard only
fixes errata in the spec, but it does let you guarantee a compiler that
accepts -std=c17 for projects that insist on using it. I guess it's
theoretically possible that following the earlier erroneous spec
results in compiler behaviour that breaks some code, too, though
implementors usually won't follow the letter of the spec if it says to
do something clearly broken.