cregistry: Avoid segfault on extended result codes
When an SQLite call returned one of the extended result codes (see
https://sqlite.org/rescode.html#extended_result_code_list), the existing
switch/case did not handle that, errPtr remained uninitialized, and
following code crashed with a segmentation fault.

Avoid this by handling the extended result codes just like we're
handling SQLITE_ERROR. Additionally, default-initialize the errPtr
structure to avoid crashes in the future.

This manifested itself when attempting to uninstall ports from
a database that returned SQLITE_CORRUPT_INDEX (779).