tcl, tclx: Update to 8.6.12 and 8.4.4
This brings us back to state of the art Tcl. Note that the only
adaptions done together with this commit are in vendor/Makefile.in. The
code in these directories itself is unpatched. Patches to fix various
problems will be re-applied in separate commits for better visibility.

Additionally, this commit will not build: Tcl 8.6 introduces a try
statement that's incompatible with the re-implementation is currently
shipping, and some changes in configure scripts require adaptions in our
buildsystem. Those will also be done in a separate commit.