m4: Don't pass --srcdir to child projects
While passing --srcdir is in general a good idea, it no longer works for
us with Tcl 8.6, since that by itself ships a number of child packages
that break when invoked with our --srcdir (and Tcl 8.6's build system
does not filter the incorrect --srcdir passed from above).

Since we'd only need this for out of source builds, which don't work
anyway, just drop it.