macports1.0: Replace and rename try/catch
Replace the old try implementation with a BSD-licensed one from tcllib's
try module, and rename it to macports_try. Extend the original tcllib
implementation with our -pass_signal flag.

Note that this commit does not include the required adaptions to users
of the old try implementation for reviewability purposes. These changes
will arrive in a follow-up commit.