Fix Linux builds with libmd installed
The standalone libmd implementation as used on Linux includes a header
called "sha2.h" which is supposed to be relative to its own include
directory.

Unfortunately, MacPorts ships its own SHA2 implementation in pextlib (in
case neither CommonCrypto, nor libmd nor OpenSSL-like implementations
are available), which also uses the "sha2.h" header file name. This
seems to take precedence and both header files are not compatible, so
the build fails.

Move the standalone SHA2 implementation to src/compat/sha2 and update
the references.

Additionally, SHA1_File is almost never a proper function, but most
often defined as a macro set to the real function. In any case, checking
for it via AC_CHECK_FUNCS() is wrong, so switch to AC_CHECK_DECLS() and
use the updated checking macros.

Fixes: https://trac.macports.org/ticket/58060