Check SQLite runtime version
Check sqlite3_libversion_number() directly at runtime where it's not
too unwieldy to do so. In the other cases, still check at build time,
but using a new macro MP_SQLITE_VERSION, which is the lesser of
the SQLITE_VERSION_NUMBER and sqlite3_libversion_number() values found
at configure time. It's not binary portability, but it should be enough
for the case where selfupdate has to build against an SDK that doesn't
match the OS version it's building on.

Fixes: https://trac.macports.org/ticket/62929