Further adopt registry2 API
Marking where registry entries could be closed, but not actually doing
it yet in most cases, because it's all too easy to close an entry that
is still in use somewhere else. We may need some form of reference
counting to make it safe to close entries in the general case.