Fix values returned by portconfigure::get_min_gcc
These need to match the final segment of a configure.compiler value.
For example, macports-gcc-4.8 yields a version of 4.8, and
macports-gcc-7 yields 7. So 8.1 would not match macports-gcc-8 even
though the gcc8 port in fact provides a newer version than that. Also,
vercmp treats n.0 as greater than n, not equal.

Fixes: https://trac.macports.org/ticket/63281