Add opt-in Altivec (ppc7400) architecture
A number of PowerPC ports bake in Altivec instructions either
unconditionally or via run-time hardware detection (see ffmpeg and
libjpeg-turbo); this either creates unusable builds on non-Altivec (G3)
machines, or inconsistent builds across hardware.

To supply G4/G5 users with Altivec builds without ruining the party
for G3 users, introduce a new ppc7400 architecture. This does not
affect any existing ports; only ports that specifically want to offer
both Altivec and non-Altivec builds need to declare "ppc7400" as a
distinct and supported architecture, e.g.

    supported_archs ppc ppc7400 i386

Once declared, the Portfile can branch on the build_arch, e.g.

    if {${configure.build_arch} eq "ppc7400"} {
        configure.flags-append  --enable-altivec
    }

ppc7400 is opt-in in the sense that the default supported_archs list
does not include it, so this should not complicate existing builds.
Existing "ppc" builds can continue to include Altivec instructions on
the sly, at least until the Portfiles are updated by disgruntled G3
users. It is entirely up to the Portfile to supply appropriate configure
and compiler flags to enable or disable Altivec. (In the future, it may
be worth revisiting this behavior if GCC is capable of generating
high-quality Altivec from unmodified C code.)

The demotion mechanism is used so that G4/G5 machines will prefer
ppc7400 when it is available, otherwise falling back to regular ppc.