Add 'add' subcommand to port select
This can be used to create user-specific, adhoc select groups. The
syntax is

    port select --add <group> <target> <name>:<link> [<name>:<link>...]

For example, consider clangd which is installed in the clang port.
clangd is installed to bin/clangd-mp-X (for some LLVM version X). If a
user wants to use this version as simply bin/clangd they can use

    port select --add clangd bin/clangd clang-mp-13:bin/clangd-mp-13

The user can install any number of alternatives, but at least one is
required.