pextlib1.0: Cache trace lookup results for speed
99% of all time spent in process_line() in tracelib.c happens in
dep_check(). In dep_check(), time is spent as follows:

- 15-20 % is spent determining the filesystem's case sensitivity
- 50 % of the time is spent querying the port that provides the given
  path
- 25 % of the time is used to translate the port ID into a port name
- ~8 % are in the binary search of the port name in the list of
  dependencies

In a test build of the yubico-pam port, 13884 iterations of dep_check()
were run. Among those, there are only a little more than 3000 unique
paths. This means that we can get a cache hit rate of about 78 % for
a cache that stores previous lookup results.

This affects the time spent in dep_check() during a test build of
yubico-pam as follows (all numbers in return values of
mach_absolute_time()):

without cache:
  2131768084 total
  14188 calls
  ~ 150251 per call
with cache:
  1054866053 total
  14188 calls
  ~ 74349 per call

Caching thus reduces the duration by 50.5 %.