port reclaim: only keep current versions' distfiles
The idea of keeping some distfiles is that they may be needed in
future when upgrading the ports, since distfiles stay the same across
revision bumps and sometimes even when the version increases. So the
files that should be kept are those that are needed by the version of
each installed port that is currently available in the ports tree, not
those that were needed to build the installed version(s).

(cherry picked from commit cd38df962d844f5adbdbfcf8cd434a50ecfb6eb4)