port reclaim: persistently cache distfile info
The cache is simply a Tcl array mapping each port name + variants to a
dist_subdir and a list of files, plus a fingerprint (consisting of the
version, revision and Portfile sha256 hash) so the entries can be
refreshed when the ports change.

The cache is also invalidated if the global variations (i.e. from
variants.conf) change. It may be possible to avoid that with a bit of
extra cleverness.