Drop privileges for all VCS operations in mportsync
Recent git versions will by default refuse to operate on directories
that are not owned by the current user. Always dropping privileges is
probably a safer approach in general anyway.

The privilege dropping and restoring code is split out into new procs,
and called such that the checks for whether a directory is a VCS repo
are done as the dir's owner, as well as the VCS update itself.