portindex: mutex wrap reading old PortIndex
Reading the same file from multiple threads is safe in itself, but it
appears that Tcl's seek command simply calls lseek() on the underlying
fd and does not save the current offset in per-interpreter data. The
offset could thus be changed by another thread between the seek and
gets/read calls.