Add built-in test phase for all ports
For now, there's just one test, which checks that the files in the
destroot have the archs that the port is supposed to be building for.
More can be added in future.

Running 'port test' on a port that doesn't set 'test.run yes' will no
longer error, but will run the built-in tests only. Ports that do set
test.run will run both the built-in tests and the ones defined by the
Portfile.

Closes: https://trac.macports.org/ticket/45010