lint: warn about missing secure checksums by default
A warning about "missing size" displayed in PRs is all too often extra
friction for contributors, when in fact a size isn't needed at all and
is the least secure of all checksum types. There was one proposal to do
something with the file size in certain specific cases (ticket 55502),
but in 5 years it has not been implemented.