src: Support requested variants in migration code
Stop the separation into negative and positive variants and add a field
for requested variants. This simplifies the code in a few places, since
no string concatenation is required at runtime to obtain the full
variant string.

Switch to restoring requested variants only by default, since this is
also the default behavior we would do for upgrading.