_upgrade: set newly installed ports' variants better
Rather than inheriting all the variants of the depending port in this
case, apply those requested on the command line and in variants.conf,
adding +universal if required to get archs compatible with those of the
depending port. This also adds error reporting in this case if the
archs cannot be made compatible.

Closes: https://trac.macports.org/ticket/67637