Add option to let mportdepends return dep list
Using this in mportexec allows it to be called multiple times without
closing all open ports in between. It also saves a call to
dlist_append_dependents.

Clients that don't pass this arg will continue to work as before,
relying on the deps being added to the global open_mports list.