macports/restore: Fix dependency calculation
Improve dependency calculation to take the requested variants from the
snapshot into account and fix it so that all dependencies that are
needed on the current system are considered. Because a new OS may have
to compile from source for a while, these dependencies can significantly
differ from the previously installed packages.

Note that this code does attempt to deal with alternative providers for
certain functionality (such as the certsync port as an alternative for
curl-ca-bundle, or -devel ports that can be installed in place of
a non-devel port), but this code does not yet seem to work correctly.
More testing is needed.

Additionally, there should be code to deal with build failures so that
ports whose dependencies failed to compile are skipped with a useful
message. At the moment, the attempt to install the dependent port
instead prompts to re-install the dependency, which (a) interrupts the
automated flow because it waits for a user prompt and (b) is pointless,
because the dependency will just fail to build again.

Improve the output formatting so that it matches MacPorts' style by
using $macports::ui_prefix where appropriate.

What's missing for this to be ready for release is:
 - A check at the end that compares the installed requested ports with
   the requested ports in the snapshot and informs the user if ports are
   missing (and explains why they are missing).
 - Help output that instructs the user how to re-attempt the migration
   if it fails (e.g., by running port restore --last)
 - Manpages and documentation
 - Some usability improvements for snapshot management. There is
   currently no way to list all snapshots or delete old snapshots.