macports/restore: Fix alternative providers
Fix the handling of ports in the snapshot that would conflict with
a dependency of a requested port by assuming that the port in the
snapshot is an alternative provider for the dependency it conflicts
with. This should fix restore for users that have certsync installed
instead of curl-ca-bundle or, for example, cmake-devel instead of cmake.

The previous approach to attempt to solve this was to patch the
dependency graph after the dependencies have been added and replacing
them with the alternative provider. Fixing this would have required to
also remove the dependency and its otherwise unneeded transitive
dependencies from the graph. It's much simpler to instead never add the
unwanted dependencies instead, which this commit implements.