macports/restore: Add progress f/dep calculation
Dependency calculation for large installations can take a few minutes.
In my personal test sample of 88 requested ports, it took about
3 minutes. Since we insert dependencies at the beginning of the
worklist, we can use processing of a requested port and the total number
of requested ports as a progress indicator.

This commit draws a progress bar based on the number of requested ports
whose dependencies have been successfully computed. This isn't perfect,
because the various steps will not take the same amount of time, and the
last port will be reported as 100 % while its dependencies are being
processed, but it's better then just printing no progress or one that
does not really allow estimating how long the process will take.