macports/restore: Use snapshot::diff to display summary
At the end of migration, show a summary of which ports failed to
install, or which (requested) ports were installed, for example, with
changed variants due to a change in default variants.

This message should eventually be extended with information how to
re-attempt migration by re-running 'port restore' as well a pointer on
how to run 'port snapshot --diff <id> --all' to see all differences.

Additionally, the message should recommend running cleanup steps such as
reclaim and uninstall inactive.