selfupdate: report number of outdated ports
Change the updatestatusvar into a dict, and return extra status
information in it: whether the ports tree was synced, and whether
portindex still needs to be run. Use this information to print messages
referring to port(1) commands in action_selfupdate instead of in
selfupdate::main, where they may not have been appropriate if using a
different client.

Closes: https://trac.macports.org/ticket/33820