macports.tcl: Fix regression in 5b6a134d6
5b6a134d66000ca1549ea218850307ef9ae37943 caused dependencies to be
installed that were not actually required when -s was used. For example,
with iperf (and thus its runtime dependencies) and cmake (its only
build-time dependency) installed, running sudo port -st install
--unrequested iperf would start to install gperf and cmake-bootstrap,
even though they were not required.

It seems this happened because MacPorts did not longer correctly detect
that the build dependency was actually installed and proceeded to
install and upgrade all its dependencies.

Thanks to halabund and Schamschula for the report.