More portlist improvements
More use of dict, including rewrites of opIntersection and opComplement
which should perform considerably better.

No more use of regex_pat_sanitize in portlist1.0, so it moves back to
port.tcl.

Bug fix for version comparison in portlist_compare. It should be
correct more often now, but it can't always be correct if there are
underscores in the version because of inherent ambiguity with the
combined version and revision.