portuninstall improvements
Use dict for options, portinfo, and variations.

Don't use portlist_sortint since not only was it being passed a list in
the wrong format, but it's defined by port.tcl and thus unlikely to
exist when this API is used by other clients. Define a comparison proc
to use with lsort instead.

Use fully-qualified names for infrequently used globals.