Allow accessing options using qualified names
Trace procs are only told the name that a variable was accessed as, not
the name used when adding the trace. So use 'namespace tail' to remove
any qualifiers before looking up the option name in option_procs or
option_defaults.

It will still cause an error to access a traced option via a different
name entirely (e.g. with upvar).