dlist_match_multi: support better comparison
The options and variations for an mport are lists of key/value pairs
with no guaranteed order. To compare them correctly requires a function
more sophisticated than a simple string comparison.

Adding a new arg to dlist_match_multi to allowing specifying a
comparison function to use for any of the keys, and a new proc
called dictequal to use with it in mportopen and mportdepends.

(cherry picked from commit d2592abd50eddd25347ecf2b92e9a6445be6a661)