registry2.0: Enable progress bar for activation
This progress bar runs in three phases, all of which spend 33.3% of the
progress:

(1) Extract the archive. Uses the -callback option for the system
pextlib command to count the lines printed by the extraction progress.
Since the archive will contain extra metadata files (+PORTFILE,
+CONTENTS, +DESC, +STATE, +COMMENT) and directories this step will
produce slightly more progress than we anticipate, which we fix at the
beginning of step 2.
(2) Check for file conflicts. Start by resetting the progress to exactly
33.3%, then loop over all files in the port once to identify potential
conflicts or ports that need to be deactivated first. This slightly
increases the total number of steps because it adds directories to be
processed in phase 3.
(3) Activate the files and directories.