Add a `curl version` Tcl command to Pextlib.
This command can be useful for debugging connectivity issues  when
Pextlib becomes upgradeable via a dedicated port and uses a libcurl
that is also installed through its own dedicated port.

The command returns a Tcl dict containing information about the
libcurl version that Pextlib was built against as well as the
libcurl it is currently running against; a simplified equivalent
of `curl --version`.
Obtaining a list of names of available feature is possible only
since curl v7.87.0 so this information will probably be replaced
with a simple feature mask in the Pextlib build installed by the
MacPorts installer on most older Darwin versions.

See trac ticket #51516