Use clonefile for port images, and detect hardlinks
New portimage_mode config option controls whether the extracted
directory, the archive, or both are kept for port images. Default is
directory if the filesystem that the software directory is on supports
cloning. Otherwise the default is archive.

In directory mode, downloaded archives are extracted only once.
Subsequent activations just clone the files. If activation is across
devices for some files, fall back to copying for those files.

In all cases, hard links are now recreated correctly whenever possible
even if the activated location is on a different device than the port
image.

Closes: https://trac.macports.org/ticket/69555