snapshot: Add JSON import functionality
This completes the export & import functionality. I changed the import
to read from a file rather than stdin. For exporting, writing to stdout
is useful because it allows the export to run as root, but the resulting
file to be owned by the calling user, but for the import this isn't
a concern.