record applied patchfiles in the statefile
Adds a "patch: <patchfile>" line to the statefile and verify this record
before applying each subsequent patch. This is a very useful mechanism
during port maintenance when some patches have to be rebased; it allows
repeated application of `port patch` (or "higher") where only the
failing patch has to be reverted and already applied patches are skipped
automatically.

Also adds default fd argument to check_statefile and write_statefile in
portutil.tcl , so that they can be called externally without having to
reopen the statefile. In that case they will check/write to the current
statefile descriptor `$target_state_fd`. This will allow to call them
safely from the patch_main override procedure in the muniversal-1.1 PG.