Give the user some control over progressbar invasiveness
Progress reporting is useful but also has its drawbacks, like writing
to the calling terminal when it's not in the appropriate state, or
slowing down the process being reported on. This commit addresses both
situations:

1. a set of 3 configuration variables is introduced with which the user
can control (via macports.conf) after how long the progressbar should
appear, if it should also show indeterminate progress (aka "we're busy")
and what operations should report progress in addition to the traditional
ones (fetch, build, destroot). The defaults for these are the current
hardwired settings. The "also_for" setting is designed to be a list that
currently only accepts "de/activation" for the recent progress reporting
added to `port activate` and `port deactivate`. (Which causes a significant
increase in duration of those operations.)

2. a Pextlib function is introduced to determine whether `port` runs as a
foreground process connected to a terminal. This function is called to
check the current state before doing any actual progressbar output. It
thus becomes possible to push a long-running operation (say, a build) to
the background and do other work in the same terminal, without getting
frequent terminal and most likely illegible output.