Adjust preferred_hosts implementation
Use 0 as the reported ping time for hosts in preferred_hosts, since the
previously used value of 1 is not guaranteed to be lower than hosts
that respond extremely quickly. 0 was previously used to indicate no
entry in the cache, so switch to using the empty string for that.

Closes: https://github.com/macports/macports-base/pull/351
(cherry picked from commit ed1bda289afb5caab12e21840c72779aef21bab5)