pextlib1.0: re-copy to SIP workaround if pextlib changed
When we modify the copying algorithm for the trace mode SIP workaround,
re-copy the files if the mtime didn't change since they were initially
copied, but their ctime is older than that of the Pextlib.dylib (or
darwintrace.dylib) that contains the copying code.

This allows us to modify the copying algorithm, which will then cause
all previously copied files to be considered outdated.