darwintracelib1.0: Fix tests
Invoking [setup] enables the tracing sandbox, which means that binaries
started after that must be started via system or must not be arm64e
system binaries (which are untraceable unmodified and will crash with
DYLD_INSERT_LIBRARIES set).

The reverse applies for [expect] in the cleanup phase.

Use system for invocations of system utilities that should be traced
when invoked after [setup]. Move invocations that don't need to be
traced after [expect] in cleanup.

Add infrastructure required to use system (e.g., allow access to the
shell and provide `ui_info` and `ui_debug`).