Avoid repeatedly building tcllibc and critcl
These build as part of the install command and don't seem to have
checks to skip this if nothing has been updated. So create a file to
indicate that they have already been destrooted.

Note that this means that manual cleaning or removing the .stamp file
is needed after changing the source of these vendored packages, which
should however hopefully be infrequent.