pextlib1.0: Fix corner case in trace mode re-signing
In particular, I've seen this failure mode when building the openjdk11
port in trace mode. I don't know why this happens, and it feels like
a bug in macOS that I can successfully posix_spawn a process, get its
pid, waitpid on that pid, and then be told that my process does not have
children, but as long as the re-signing works (which it seems to),
I don't care.

(cherry picked from commit 6dda4cc6c08176cf497dd1417189480830a357a1)