postflight: check RUNUSR attributes
As done in the Makefile in ad2f4ab. Theoretically this problem
shouldn't happen here, as RUNUSR is always 'macports', but there have
been occasional reports of strange things happening to the macports
user or other accounts getting renamed to 'macports' (with Migration
Assistant being one of the likely culprits), so it's probably best not
to mess with the user if it doesn't look right.