Rework extract options
All extract.* options are now defined and handled in portextract.tcl,
rather than some being in portfetch.tcl.

The extract.cmd, extract.pre_args, extract.post_args, and extract.suffix
options now have dynamic defaults based on which of the use_* options
are set (if any).

Setting one of the use_* options to a true value now resets all the
other ones.

Extract dependencies are now added in a callback so that they can be
based on the final values of the relevant options.

Closes: https://trac.macports.org/ticket/64960